EE6008_INSERT_COMMENTS

+++ online_library/src/main/java/com/onlineLibrary/aspect/AutoFillAspect.java
@@ -43,15 +43,15 @@ public class AutoFillAspect {
         Object entity = args[0];
         //准备赋值数据
         LocalDateTime now = LocalDateTime.now();
-        Long currentId = BaseContext.getCurrentId();
+        Integer currentId = BaseContext.getCurrentId();
         //根据类型赋值 插入赋值4条 更新赋值两条
         if (operationType == OperationType.INSERT) {
             try {
                 //拦截对象添加方法
                 Method setCreateTime = entity.getClass().getDeclaredMethod(AutoFillConstant.SET_CREATE_TIME,LocalDateTime.class);
-                Method setCreateUser = entity.getClass().getDeclaredMethod(AutoFillConstant.SET_CREATE_USER,Long.class);
+                Method setCreateUser = entity.getClass().getDeclaredMethod(AutoFillConstant.SET_CREATE_USER,Integer.class);
                 Method setUpdateTime = entity.getClass().getDeclaredMethod(AutoFillConstant.SET_UPDATE_TIME,LocalDateTime.class);
-                Method setUpdateUser = entity.getClass().getDeclaredMethod(AutoFillConstant.SET_UPDATE_USER,Long.class);
+                Method setUpdateUser = entity.getClass().getDeclaredMethod(AutoFillConstant.SET_UPDATE_USER,Integer.class);
                 //赋值
                 setCreateTime.invoke(entity, now);
                 setCreateUser.invoke(entity, currentId);
@@ -64,13 +64,12 @@ public class AutoFillAspect {
             try {
                 //拦截对象添加方法
                 Method setUpdateTime = entity.getClass().getDeclaredMethod(AutoFillConstant.SET_UPDATE_TIME,LocalDateTime.class);
-                Method setUpdateUser = entity.getClass().getDeclaredMethod(AutoFillConstant.SET_UPDATE_USER,Long.class);
+                Method setUpdateUser = entity.getClass().getDeclaredMethod(AutoFillConstant.SET_UPDATE_USER,Integer.class);
                 //赋值
                 setUpdateTime.invoke(entity, now);
                 setUpdateUser.invoke(entity, currentId);
             } catch (Exception e) {
-                throw new RuntimeException(e);
-            }
+                throw new RuntimeException(e);}
         }
